labs/lab01/requirements.txt
appdirs==1.4.4
astroid==2.5
black==20.8b1
click==7.1.2
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
pathspec==0.8.1
protobuf==3.15.3
pylint==2.7.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
